pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swot-toolkit"
version = "0.1.0"
description = "A Python package for SWOT satellite data utilities"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy",
    "xarray",
    "geopandas",
    "earthaccess",
    "h5netcdf",
    "matplotlib",
    "shapely",
    "fastkml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
    "jupyter",
    "ipywidgets",
]

[project.urls]
Homepage = "https://github.com/yourusername/swot-toolkit"
Repository = "https://github.com/yourusername/swot-toolkit.git"
Issues = "https://github.com/yourusername/swot-toolkit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/swot_toolkit"]

[tool.ruff]
# Enable all checks by default
select = ["ALL"]
ignore = ["EXE002"]

line-length = 99

# assume python 3.10
target-version = "py310"

[tool.pyright]
typeCheckingMode = "strict"
reportUnknownMemberType = "none"
ignore = ["**/nbs"]
reportMissingImports = false
reportMissingTypeStubs = false


# line-length = 88
# target-version = "py39"
# select = [
#     "E",  # pycodestyle errors
#     "W",  # pycodestyle warnings
#     "F",  # pyflakes
#     "I",  # isort
#     "B",  # flake8-bugbear
#     "C4", # flake8-comprehensions
#     "UP", # pyupgrade 
# ]
# ignore = [
#     "E501", # line too long, handled by black
# ]

# [tool.ruff.format]
# quote-style = "double"
# indent-style = "space"
# skip-magic-trailing-comma = false
# line-ending = "auto"

# [tool.ruff.lint.isort]
# known-first-party = ["swot_toolkit"]

# [tool.pytest.ini_options]
# testpaths = ["tests"]
# python_files = ["test_*.py"]
# python_classes = ["Test*"]
# python_functions = ["test_*"]
# addopts = [
#     "--strict-markers",
#     "--strict-config",
#     "--cov=swot_toolkit",
#     "--cov-report=term-missing",
# ]